style(rocprofiler-systems): enable phase-1 clang-tidy readability checks
Scope the .clang-tidy readability-* set to a low-risk first pass and apply
the corresponding fixes:

- Fix the inherited no-op '-readability-implicit-bool-cast' by disabling the
  real 'readability-implicit-bool-conversion' check (removes the largest and
  least valuable source of churn).
- Defer signature-changing and non-auto-fixable checks to a later phase
  (convert-member-functions-to-static, non-const-parameter,
  make-member-function-const, avoid-nested-conditional-operator,
  suspicious-call-argument, avoid-unconditional-preprocessor-if).
- Apply 191 auto-fixes across 36 files for the enabled local checks
  (container-contains, avoid-return-with-void-value, use-std-min-max,
  qualified-auto, container-size-empty, redundant-casting, duplicate-include,
  redundant-control-flow); reformatted with clang-format-18.

Wire up enforcement:
- ROCPROFSYS_USE_CLANG_TIDY now runs clang-tidy with -warnings-as-errors=* so
  any finding fails the build (only the project's own source/ targets are
  analyzed; third-party deps, Python bindings and examples are excluded).
- Add a blocking clang-tidy job to the rocprofiler-systems formatting workflow
  that configures + builds with ROCPROFSYS_USE_CLANG_TIDY=ON in the CI image.

+++ projects/rocprofiler-systems/source/lib/core/perf.cpp
@@ -173,7 +173,7 @@ config_overflow_sampling(struct perf_event_attr& _pe, std::string_view _event,
         }
         case PERF_TYPE_HW_CACHE:
         {
-            _pe.config = static_cast<int>(perf::get_hw_cache_config(_event));
+            _pe.config = perf::get_hw_cache_config(_event);
             break;
         }
         case PERF_TYPE_BREAKPOINT:
